create NodeMetaManager
re-label NodeMetaFormatter headers
